requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
boto3==1.34.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pymongo==4.6.0